ai_assistant/requirements.txt
flask
flask-session 
flask-cors
google-generativeai 
PyPDF2 
markdown
python-dotenv